+++ contracts/DCAFactory/DCAFactoryParameters.sol
@@ -5,43 +5,43 @@ import '@openzeppelin/contracts/utils/structs/EnumerableSet.sol';
 
 interface IDCAFactoryParameters {
   event FeeRecipientSet(address _feeRecipient);
-  event FeeSet(uint256 _feeSet);
-  event SwapIntervalsAllowed(uint256[] _swapIntervals);
-  event SwapIntervalsForbidden(uint256[] _swapIntervals);
+  event FeeSet(uint32 _feeSet);
+  event SwapIntervalsAllowed(uint32[] _swapIntervals);
+  event SwapIntervalsForbidden(uint32[] _swapIntervals);
 
   /* Public getters */
 
   function feeRecipient() external view returns (address);
 
-  function fee() external view returns (uint256);
+  function fee() external view returns (uint32);
 
   // solhint-disable-next-line func-name-mixedcase
-  function FEE_PRECISION() external view returns (uint256);
+  function FEE_PRECISION() external view returns (uint24);
 
   // solhint-disable-next-line func-name-mixedcase
-  function MAX_FEE() external view returns (uint256);
+  function MAX_FEE() external view returns (uint32);
 
-  function allowedSwapIntervals() external view returns (uint256[] memory __allowedSwapIntervals);
+  function allowedSwapIntervals() external view returns (uint32[] memory __allowedSwapIntervals); // uint32 is enough for 100 years
 
-  function isSwapIntervalAllowed(uint256 _swapInterval) external view returns (bool);
+  function isSwapIntervalAllowed(uint32 _swapInterval) external view returns (bool);
 
   /* Public setters */
   function setFeeRecipient(address _feeRecipient) external;
 
-  function setFee(uint256 _fee) external;
+  function setFee(uint32 _fee) external;
 
-  function addSwapIntervalsToAllowedList(uint256[] calldata _swapIntervals) external;
+  function addSwapIntervalsToAllowedList(uint32[] calldata _swapIntervals) external;
 
-  function removeSwapIntervalsFromAllowedList(uint256[] calldata _swapIntervals) external;
+  function removeSwapIntervalsFromAllowedList(uint32[] calldata _swapIntervals) external;
 }
 
 abstract contract DCAFactoryParameters is IDCAFactoryParameters {
   using EnumerableSet for EnumerableSet.UintSet;
 
   address public override feeRecipient;
-  uint256 public override fee = 3000; // 0.3%
-  uint256 public constant override FEE_PRECISION = 10000;
-  uint256 public constant override MAX_FEE = 10 * FEE_PRECISION; // 10%
+  uint32 public override fee = 3000; // 0.3%
+  uint24 public constant override FEE_PRECISION = 10000;
+  uint32 public constant override MAX_FEE = 10 * FEE_PRECISION; // 10%
   EnumerableSet.UintSet internal _allowedSwapIntervals;
 
   constructor(address _feeRecipient) {
@@ -54,13 +54,13 @@ abstract contract DCAFactoryParameters is IDCAFactoryParameters {
     emit FeeRecipientSet(_feeRecipient);
   }
 
-  function _setFee(uint256 _fee) internal {
+  function _setFee(uint32 _fee) internal {
     require(_fee <= MAX_FEE, 'DCAFactory: fee too high');
     fee = _fee;
     emit FeeSet(_fee);
   }
 
-  function _addSwapIntervalsToAllowedList(uint256[] calldata _swapIntervals) internal {
+  function _addSwapIntervalsToAllowedList(uint32[] calldata _swapIntervals) internal {
     for (uint256 i = 0; i < _swapIntervals.length; i++) {
       require(_swapIntervals[i] > 0, 'DCAFactory: zero interval');
       require(!isSwapIntervalAllowed(_swapIntervals[i]), 'DCAFactory: allowed swap interval');
@@ -69,7 +69,7 @@ abstract contract DCAFactoryParameters is IDCAFactoryParameters {
     emit SwapIntervalsAllowed(_swapIntervals);
   }
 
-  function _removeSwapIntervalsFromAllowedList(uint256[] calldata _swapIntervals) internal {
+  function _removeSwapIntervalsFromAllowedList(uint32[] calldata _swapIntervals) internal {
     for (uint256 i = 0; i < _swapIntervals.length; i++) {
       require(isSwapIntervalAllowed(_swapIntervals[i]), 'DCAFactory: swap interval not allowed');
       _allowedSwapIntervals.remove(_swapIntervals[i]);
@@ -77,15 +77,15 @@ abstract contract DCAFactoryParameters is IDCAFactoryParameters {
     emit SwapIntervalsForbidden(_swapIntervals);
   }
 
-  function allowedSwapIntervals() external view override returns (uint256[] memory __allowedSwapIntervals) {
+  function allowedSwapIntervals() external view override returns (uint32[] memory __allowedSwapIntervals) {
     uint256 _allowedSwapIntervalsLength = _allowedSwapIntervals.length();
-    __allowedSwapIntervals = new uint256[](_allowedSwapIntervalsLength);
+    __allowedSwapIntervals = new uint32[](_allowedSwapIntervalsLength);
     for (uint256 i = 0; i < _allowedSwapIntervalsLength; i++) {
-      __allowedSwapIntervals[i] = _allowedSwapIntervals.at(i);
+      __allowedSwapIntervals[i] = uint32(_allowedSwapIntervals.at(i));
     }
   }
 
-  function isSwapIntervalAllowed(uint256 _swapInterval) public view override returns (bool) {
+  function isSwapIntervalAllowed(uint32 _swapInterval) public view override returns (bool) {
     return _allowedSwapIntervals.contains(_swapInterval);
   }
 }
